+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [1.0.1]
+- Adds `content` param to pass custom widget in your content.Thanks [robisaks](https://github.com/robisaks)
+- Migrate paraem `context` to `show` method.
+
 ## [1.0.0]
 
 - Migrate to Null Safety [PR #09](https://github.com/RafaelBarbosatec/achievement_view_flutter/pull/9).